revert maintenance view
