+++ react/src/index.tsx
@@ -5,16 +5,6 @@ const DefaultProviders = React.lazy(
   () => import("./components/DefaultProviders")
 );
 const Information = React.lazy(() => import("./components/Information"));
-const Maintenance = React.lazy(() => import("./components/Maintenance"));
-
-customElements.define(
-  "backend-ai-react-maintenance",
-  reactToWebComponent((props) => (
-    <DefaultProviders {...props}>
-      <Maintenance />
-    </DefaultProviders>
-  ))
-);
 
 customElements.define(
   "backend-ai-react-information",
